requirements.txt
aiogram==3.2.0
aiosqlite==0.19.0
python-dotenv==1.0.0
openai>=1.0.0
httpx>=0.24.0
